Add GitHub fork me ribbon
This was requested in the Slack chat. I grabbed the code from [GitHub
Ribbons][1] using the same color as the old site.

One caveat was that the z-index was lower then the header nav. So I
added the `z-index: 9999` to the style tag.

I didn't want to make this a CSS rule because I was trying to stay as
close to the original copy/paste code as I could.

[1]: https://github.com/blog/273-github-ribbons