Add meetup-events component
This component has the added extra of managing the network fetch of data
and appending the meetup-event-card for each one.

Meetup.com does not offer CORS unless the client authorizes through
OAuth. Since it was silly to expose a member's OAuth token just to how
the events and it was silly to ask a viewer to authenticate we are
forced to use JSONP.

Originally I wrote all of the code using jQuery. I realized that
web components are more of a standard then jQuery and that most
everything could be managed with Vanilla JS. My hope was that the code
could be packaged into a lib for other groups' websites. And so I
stripped out all the jQuery.

Testing JSONP is also a fools errand so I pushed all the logic for JSONP
into one function that can be mocked out.