Wait for HTML imports before running test suite
This was an odd one. Seems HTML imports are parsed asynchronously but in
order. In other words they don't block other scripts on the page. This
didn't seem to matter in Chrome but in Firefox it broke ALL the tests!

Instead we use the imports onload to register that it is ready. When all
the imports are done then we run the tests.