Fix tests in chrome
Sigh, This is getting ridiculous. The problems seem to be that the folks
who wrote the webcomponents polyfill either neglected to mention browser
differences or the browsers are being stupid at following specs.

* Chrome blocks on HTML Imports
* Firefox does not

So the fix that worked for Firefox cause chrome to execute mocha.run
before the test cases were loaded. (╯°□°）╯︵┻━┻

Changing the order fixes it. I swear I've now tested this in

* Firefox
* Chrome
* Safari

Really sorry for the multiple pull requests.