Add Security PGP public key
This is the signing key for security.txt and the key that researchers
can encrypt to if they feel the need to keep communication secret.

The private (secret) is stored in Keybase.io to the newhavenio.admins
team where only a select few members have access and that membership
list can be adjusted over time.