Add all-contributors support
https://github.com/kentcdodds/all-contributors

To add a contributor simply run:

    # Add new contributor <username>, who made a contribution of type <contribution>
    yarn run contributors:add <username> <contribution>
    # Example:
    yarn run contributors:add jfmengels code,doc

Where username is the user's GitHub or Gitlab username, and contribution
is a ,-separated list of ways to contribute, [listed
here](https://github.com/kentcdodds/all-contributors#emoji-key).

This is a specification for recognizing contributors to an open source
project in a way that rewards each and every contribution, not just
code.

The basic idea is this:

> Use the project README (or other prominent public documentation page
> in the project) to recognize the contributions of members of the
> project community.

People are giving of themselves and their free time to contribute to
open source projects in so many ways. It can be a real time sink
sometimes and so they should be praised for all their contributions
(code or not).